requirements.txt
django==2.1.2
djangorestframework==3.8.2
django-cors-headers==2.4.0
djangorestframework_simplejwt==3.2.3
psycopg2-binary